+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Adds label propagation for semi-supervised learning (#43)
 * Adds switch `--ignore` to all prediction-related commands, for ignoring layers (#43)
 * Adds weight decay for better regularization during training (#44)
+* Can handle missing/no-data class labels (#45)
 
 ### Bugfixes